refactor: Move Person package to Contact
