+++ src/main/java/seedu/address/logic/Messages.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import seedu.address.logic.parser.Prefix;
-import seedu.address.model.person.Person;
+import seedu.address.model.contact.Person;
 
 /**
  * Container for user visible messages.
