+++ src/main/java/seedu/address/model/AddressBook.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.contact.Person;
+import seedu.address.model.contact.UniquePersonList;
 
 /**
  * Wraps all data at the address-book level
