+++ src/main/java/seedu/address/model/contact/exceptions/PersonNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.contact.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified person.
