+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -17,8 +17,8 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
-import seedu.address.model.person.Person;
+import seedu.address.model.contact.NameContainsKeywordsPredicate;
+import seedu.address.model.contact.Person;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 
 /**
