+++ src/test/java/seedu/address/model/contact/PhoneTest.java
@@ -1,12 +1,10 @@
-package seedu.address.model.person;
+package seedu.address.model.contact;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
-
-import seedu.address.model.contact.Phone;
 
 public class PhoneTest {
